Add code to deal with directories, small UI tweaks
-Add Mod Dialog-
Will not open if DDLC base folder has not been set, popup will come up notifying user to download DDLC/fix their settings.
Adding a mod copies DDLC to the new folder and opens it up in a file browser.
   ~If the new mod replaces an old one, it will delete the old directory first, so beware!

-DDLC-
Has been renamed to DDLC Mod Manager.py from ddlc.py
'Play', 'Remove/Copy/Rename Mod', and 'Open Mod Folder buttons all now work as expected, and will open a popup if a mod has not been selected.
   ~'Play' should work on both Windows and Linux, but only Windows has been tested so far. Further work will need to be done to get OSX support working.

+++ ddlcfldsel.py
@@ -8,9 +8,8 @@
 
 from PyQt5 import QtCore, QtGui, QtWidgets
 from PyQt5.QtWidgets import QFileDialog, QMessageBox
-import os
-import configparser
-import io
+import os, configparser, io
+
 
 class Ui_Fld_Sel(object):
 
@@ -19,49 +18,51 @@ class Ui_Fld_Sel(object):
         config = configparser.ConfigParser(allow_no_value=True)
         config.read('config.ini')
         gameFld = config['DEFAULT']['ddlcfolder']
-#        print(gameFld)
 
-
-
-
+        global process
+        process = Fld_Sel
         Fld_Sel.setObjectName("Fld_Sel")
-        Fld_Sel.resize(546, 170)
+        Fld_Sel.resize(442, 170)
         Fld_Sel.setMaximumSize(QtCore.QSize(16777215, 170))
         Fld_Sel.setWhatsThis("")
         self.gridLayout = QtWidgets.QGridLayout(Fld_Sel)
         self.gridLayout.setObjectName("gridLayout")
-        self.openDirectory110 = QtWidgets.QToolButton(Fld_Sel)
-        self.openDirectory110.setObjectName("openDirectory110")
-        self.gridLayout.addWidget(self.openDirectory110, 1, 3, 1, 1)
-#Open DDLC/game folder if images-110.rpa is inside
-        self.openDirectory110.clicked.connect(self.openGameFld)
-
-
-        self.chooseDirectory111 = QtWidgets.QToolButton(Fld_Sel)
-        self.chooseDirectory111.setCheckable(False)
-        self.chooseDirectory111.setToolButtonStyle(QtCore.Qt.ToolButtonIconOnly)
-        self.chooseDirectory111.setObjectName("chooseDirectory111")
-        self.gridLayout.addWidget(self.chooseDirectory111, 3, 3, 1, 1)
-#Choose folder that DDLC is stored in
-        self.chooseDirectory111.clicked.connect(self.chsFld)
-
-
-        self.folderPath111 = QtWidgets.QLineEdit(Fld_Sel)
-        self.folderPath111.setInputMask("")
-        self.folderPath111.setText("")
-        self.folderPath111.setObjectName("folderPath111")
-        self.gridLayout.addWidget(self.folderPath111, 3, 1, 1, 1)
         self.folderLabel110 = QtWidgets.QLabel(Fld_Sel)
         self.folderLabel110.setMaximumSize(QtCore.QSize(16777215, 24))
         self.folderLabel110.setObjectName("folderLabel110")
         self.gridLayout.addWidget(self.folderLabel110, 0, 1, 1, 1)
+        self.openDirectory110 = QtWidgets.QToolButton(Fld_Sel)
+        self.openDirectory110.setObjectName("openDirectory110")
+        self.gridLayout.addWidget(self.openDirectory110, 1, 3, 1, 1)
+        #Open 'DDLC\game'
+        self.openDirectory110.clicked.connect(self.openGameFld)
+
+        self.imagesPathLabel = QtWidgets.QLabel(Fld_Sel)
+        self.imagesPathLabel.setMinimumSize(QtCore.QSize(265, 0))
+        self.imagesPathLabel.setObjectName("imagesPathLabel")
+        self.gridLayout.addWidget(self.imagesPathLabel, 1, 1, 1, 1)
         self.folderLabel111 = QtWidgets.QLabel(Fld_Sel)
         self.folderLabel111.setMaximumSize(QtCore.QSize(16777215, 24))
         self.folderLabel111.setObjectName("folderLabel111")
         self.gridLayout.addWidget(self.folderLabel111, 2, 1, 1, 1)
-        self.imagesPathLabel = QtWidgets.QLabel(Fld_Sel)
-        self.imagesPathLabel.setObjectName("imagesPathLabel")
-        self.gridLayout.addWidget(self.imagesPathLabel, 1, 1, 1, 1)
+        self.folderPath111 = QtWidgets.QLineEdit(Fld_Sel)
+        self.folderPath111.setInputMask("")
+        self.folderPath111.setText("")
+        self.folderPath111.setObjectName("folderPath111")
+        self.gridLayout.addWidget(self.folderPath111, 3, 1, 1, 2)
+        self.chooseDirectory111 = QtWidgets.QToolButton(Fld_Sel)
+        self.chooseDirectory111.setObjectName("chooseDirectory111")
+        self.gridLayout.addWidget(self.chooseDirectory111, 3, 3, 1, 1)
+        #Choose DDLC directory
+        self.chooseDirectory111.clicked.connect(self.chsFld)
+
+        self.DlgBtns = QtWidgets.QDialogButtonBox(Fld_Sel)
+        self.DlgBtns.setStandardButtons(QtWidgets.QDialogButtonBox.Cancel|QtWidgets.QDialogButtonBox.Ok)
+        self.DlgBtns.setObjectName("DlgBtns")
+        self.gridLayout.addWidget(self.DlgBtns, 4, 2, 1, 1)
+        self.DlgBtns.accepted.connect(self.saveConfig)
+
+        self.DlgBtns.rejected.connect(Fld_Sel.close)
 
         self.retranslateUi(Fld_Sel)
         QtCore.QMetaObject.connectSlotsByName(Fld_Sel)
@@ -69,17 +70,14 @@ class Ui_Fld_Sel(object):
     def retranslateUi(self, Fld_Sel):
         _translate = QtCore.QCoreApplication.translate
         Fld_Sel.setWindowTitle(_translate("Fld_Sel", "Download DDLC"))
+        self.folderLabel110.setText(_translate("Fld_Sel", "DDLC v1.1.0 images.rpa Location:"))
         self.openDirectory110.setText(_translate("Fld_Sel", "..."))
-        self.chooseDirectory111.setText(_translate("Fld_Sel", "..."))
+        self.imagesPathLabel.setText(_translate("Fld_Sel", "images-110.rpa does not exist or is not in DDLC/game/"))
+        self.folderLabel111.setText(_translate("Fld_Sel", "DDLC v1.1.1 Location:"))
         self.folderPath111.setPlaceholderText(_translate("Fld_Sel", "Choose which folder v1.1.1 is located."))
-#Set it to display game folder if set in config.ini
         if gameFld != None:
             self.folderPath111.setText(gameFld)
-
-
-        self.folderLabel110.setText(_translate("Fld_Sel", "DDLC v1.1.0 images.rpa Location:"))
-        self.folderLabel111.setText(_translate("Fld_Sel", "DDLC v1.1.1 Location:"))
-        self.imagesPathLabel.setText(_translate("Fld_Sel", "images-110.rpa does not exist or is not in DDLC/game/"))
+        self.chooseDirectory111.setText(_translate("Fld_Sel", "..."))
 
     def chsFld(self):
         global gameFld
@@ -90,21 +88,25 @@ class Ui_Fld_Sel(object):
             #Update text on folder selector dialog to reflect new location
             self.folderPath111.setText(gameFld)
             #Save new game directory to config.ini
-            config = configparser.ConfigParser(allow_no_value=True)
-            config.read('config.ini')
-            config['DEFAULT']['ddlcfolder'] = gameFld
-
-            with open('config.ini', 'w') as configfile:
-                config.write(configfile)
         else:
             pass
 
+    def saveConfig(self):
+            if os.path.exists(self.folderPath111.text()):
+                config = configparser.ConfigParser(allow_no_value=True)
+                config.read('config.ini')
+                config['DEFAULT']['ddlcfolder'] = self.folderPath111.text()
+                with open('config.ini', 'w') as configfile:
+                    config.write(configfile)
+                process.hide()
+            else:
+                QMessageBox.question(None, 'Error', "Given location for DDLC is not a valid directory.", QMessageBox.Ok)
+                self.folderPath111.setText(None)
+
     def openGameFld(self):
-#        print("niceu")
         try:
             if gameFld != None:
 #Checking if a location for DDLC has been set
-#                print("niceur")
                 if os.path.join(gameFld, "game") != None:
                     os.startfile(os.path.join(gameFld, "game"))
 #Checking if DDLC installation is valid, i.e. does it have a "game" folder
@@ -113,6 +115,7 @@ class Ui_Fld_Sel(object):
             QMessageBox.question(None, 'Error', "There is no '/DDLC/game' folder or DDLC location is incorrect, reinstall DDLC.", QMessageBox.Ok)
         except NameError:
             print("No location set for DDLC")
+
 
 if __name__ == "__main__":
     import sys
